Setting up theming to save to file and use data option
