--- app/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<link rel="stylesheet" href="proschools.css">
-	<meta charset="UTF-8">
-	<title>Title</title>
-</head>
-<body>
-	<div class="fun">YES</div>
-</body>
-</html>
