--- app/proschools.css
@@ -1,25 +1,0 @@
-.true {
-  color: red;
-  display: -webkit-box;
-  display: -webkit-flex;
-  display: -ms-flexbox;
-  display: flex; }
-
-.fun {
-  color: blue; }
-
-.sleep {
-  color: red; }
-
-.true {
-  color: red;
-  display: -webkit-box;
-  display: -webkit-flex;
-  display: -ms-flexbox;
-  display: flex; }
-
-.fun {
-  color: blue; }
-
-.sleep {
-  color: red; }
