+++ app/public/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	<link rel="stylesheet" href="/portal/styles/demo/color1/red/color2/green/color3/orange/style.css">
+	<meta charset="UTF-8">
+	<title>Title</title>
+</head>
+<body>
+	<div class="fun">FUN</div>
+	<div class="true">TRUE</div>
+	<div class="sleep">SLEEP</div>
+</body>
+</html>
